{"type":"TI", "runs":"1"}

+++ Hangman.java
@@ -54,7 +54,7 @@ public class Hangman extends ConsoleProgram {
     	char user_letter_ch;
     	
     	// get a guessing word from the dictionary
-    	int high_boundary = hangmanlecxicon.getWordCount();
+    	double high_boundary = hangmanlecxicon.getWordCount();
     	int word_index = (int) rgen.nextDouble(0, high_boundary);
     	guess_word = hangmanlecxicon.getWord(word_index);
     	println("guessed word is: " + guess_word + " num of words are: " + high_boundary);
